:white_check_mark: - Move tests to lib so they only run once
